requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sentence-transformers==3.4.1
faiss-cpu==1.10.0
pydantic==2.10.6
python-dotenv==1.0.1
openai==1.61.1
httpx==0.28.1
beautifulsoup4==4.12.3
readability-lxml==0.8.1
slowapi==0.1.9
sentry-sdk==2.20.0
google-api-python-client==2.161.0
google-auth==2.38.0
pypdf==5.3.0
python-docx==1.1.2
python-pptx==1.0.2
youtube-transcript-api==0.6.3
Pillow==11.1.0
python-multipart==0.0.20
defusedxml==0.7.1
PyYAML==6.0.2
